Task 1 and 2 for Full stack ASP.NET developer (remote) position. Checked in by Justin Dong (Justin in Austin TX USA)
Task #1. You have an online store that sells books. Add a new “Author” property to “Product” entity. A store owner should be able to edit it in admin area. And it should be displayed to customers in public store.
Task #2. Implement a new widget plugin (IWidgetPlugin). This plugin should display a message on the product details page in public store (e.g. “50% discount in December”). And this message should be editable in admin area on the widget configuration page.